add docker rmi

+++ 3.SpringFrameWork/AdminUI/ClearImage.sh
@@ -3,5 +3,7 @@
 # remove the specific docker image
 
 echo "clean image"
-docker rmi $(docker images -q $1)
+if [-n docker images -q $1]
+  docker rmi $(docker images -q $1)
+fi
